+++ lib/modal_progress_hud.dart
@@ -3,40 +3,50 @@ library modal_progress_hud;
 import 'package:flutter/material.dart';
 
 class ModalProgressHUD extends StatelessWidget {
-  final Widget child;
-  final bool inAsyncCall;
+  final Widget _child;
+  final bool _inAsyncCall;
   final double opacity;
   final Color color;
-  final ProgressIndicator progressIndicator;
+  final Widget progressIndicator;
+  final Offset offset;
 
   ModalProgressHUD({
     Key key,
-    @required this.child,
-    @required this.inAsyncCall,
+    @required child,
+    @required inAsyncCall,
     this.opacity = 0.3,
     this.color = Colors.grey,
     this.progressIndicator = const CircularProgressIndicator(),
-  }) : super(key: key);
+    this.offset,
+  })  : assert(child != null),
+        assert(inAsyncCall != null),
+        _child = child,
+        _inAsyncCall = inAsyncCall,
+        super(key: key);
 
   @override
   Widget build(BuildContext context) {
     List<Widget> widgetList = new List<Widget>();
-    widgetList.add(child);
-    if (inAsyncCall) {
-      final modal = new Stack(
-        children: [
-          new Opacity(
-            opacity: opacity,
-            child: ModalBarrier(dismissible: false, color: color),
-          ),
-          new Center(
-            child: progressIndicator,
-          ),
-        ],
+    widgetList.add(_child);
+    Widget layOutProgressIndicator = new Center(child: progressIndicator);
+    if (offset != null) {
+      layOutProgressIndicator = new Positioned(
+        child: progressIndicator,
+        left: offset.dx,
+        top: offset.dy,
       );
-      widgetList.add(modal);
     }
-    return Stack(
+    if (_inAsyncCall) {
+      final modal = [
+        new Opacity(
+          opacity: opacity,
+          child: new ModalBarrier(dismissible: false, color: color),
+        ),
+        layOutProgressIndicator
+      ];
+      widgetList.addAll(modal);
+    }
+    return new Stack(
       children: widgetList,
     );
   }